+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0]
+### Added
+- Use order/properties-order to specify custom properties order
+### Changes
+- Disable order/properties-alphabetical-order
+- Disable selector-no-qualifying-type for attributes and classes
+
+
 ## [1.0.0]
 ### Added
 - Initial release, based from [stylelint-config-sass-guidelines](https://github.com/bjankord/stylelint-config-sass-guidelines)